deps(config): extract configuration to a new package
