+++ configuration/config.go
@@ -1,17 +1,22 @@
-package main
+package configuration
 
 import (
+	"fmt"
+
 	"github.com/kelseyhightower/envconfig"
 	"github.com/pkg/errors"
 )
 
 type Config struct {
-	Port int `envconfig:"PORT" default:"3000"`
+	Port int `envconfig:"PORT" `
 }
 
-func newConfig() (*Config, error) {
+func NewConfig() (*Config, error) {
 	var cfg Config
-	err := envconfig.Process("", &cfg)
+	fmt.Println(cfg)
+	fmt.Println(&cfg)
+
+	err := envconfig.Process("myapp", &cfg)
 	if err != nil {
 		return nil, errors.Wrapf(err, "fail to build config from env")
 	}
